Rework dired listing switches.
Originally, I had dired use GNU ls switches as long as system-type was
gnu/linux. This would cause issues if I were to ever use a Linux system
that does not use the GNU coreutils (like Alpine?). Now, it explicitly
checks for "GNU coreutils" in the ls --version string and falls back to
a safer set of flags if not found.