Add local ox-extra package.
ox-extra is one small part of the larger org-contrib package: a
collection of mostly unmaintained org-mode addons. I am not
particularly interested in installing a bunch of abandoned org
plugins, but ox-extra provides a useful :ignore: behavior for
headlines and is also only a single small file. Into the lisp
directory it goes!