Add consult and devil packages.
Unfortunately, these are right next to each other, so I can't split
this into two commits. Consult is set up with replacements for common
actions, now with fancy previews. I currently use the default comma as
the devil key, but I will probably move it to semicolon once the new
version with the devil-set-key function comes out.

Also hook company-mode into comint-mode, because that's in this hunk
too. Bah.