Move package setup to package.el.
Before now, setting up the package system had to be done very early in
my Emacs initialization because I had to make sure that use-package
was available and ready to use for the rest of my init. Now that Emacs
ships with use-package, the package setup can be moved to package.el
where all the other mucking about with packages happens. This makes
more sense.