Move browse-url-browser-function declaration.
This goes with the policy that the first massive setq-default should
only contain "Emacs settings" and customizations for other modes and
"Emacs applications" should be moved to the use-package declarations
further down. browse-url-browser-function technically belongs to the
browse-url "package" but I'm including it with url here because I
don't want to create a new use-package statement just for that.