Add tree-sitter setup.
This setup goes in its own treesit-config.el file that is loaded
whenever tree-sitter is available. Of course, I don't expect to be
using any Emacs without tree sitter for the time being, but it's still
nice to think ahead (?).

In order to have different tree sitter modes available, I could have
used the treesit-auto package from MELPA. However, I did not quite
like the behavior of the package when playing with it on my own. This
setup is more explicit and manual and behaves exactly to my liking as
far as I can tell.