Improve CJK fallback font setup.
The core issue that I keep running into is that the set-fontset-font
function does not reliably work at Emacs startup. In particular, it
doesn't work when Emacs is started as a daemon, which is how I
normally use Emacs. Instead, I wrap the config into a function and
call it when a frame is created.