Bind "u" to update index in mu4e.
The "U" command in mu4e's main view runs both the mail fetcher and mu
to index the result. On my machines, I already have my mail indexer
run every 15 minutes, so the extra time spent fetching mail is
annoying. This has caused me to run M-x mu4e-update-index instead of
the "U" command most of the time, so it would be nice to have that
function bound to a key.

Unfortunately, I can't figure out how to add a new entry to the mu4e
main view to show that the new binding exists, so it will just have to
exist invisibly.