Add functions for working with tags in Org Roam.
I spent some time tagging all of my Org Roam stuff, only to find out
that these tags were pretty much useless. Org Roam has a function to
add tags to a file but there are no functions to actually do anything
with that information.  So, I've learned the basics of the Roam API
and wrote some functions for it myself.

The first function is cn/org-roam-node-find-by-tag, which modifies
node search by first prompting for a tag to narrow the results.

The second function is cn/org-roam-insert-nodes-with-tag, which finds
all files with a given tag and inserts a bulleted list of links to
them. This makes it easier to quickly collect notes on a given topic.