Ignore lisp error from jumping to compile error when there are none.
That's weirdly worded, but the best I can do on a single line. I have
the compilation buffer set up to automatically jump to the first error
if compilation is not successful. However, if compilation was
successful and there are no errors, I get a lisp error from trying to
jump to a compilation error that doesn't actually exist.

It would be annoying to set up the "jump to first error" code to only
run when compilation fails (or at least it's not immediately obvious),
so instead I will just ignore the error when it happens.