Remove exec-path-from-shell.
exec-path-from-shell is supposed to populate exec-path with your
shell's PATH. However, I have been playing with alternative shells
like fish and the package really does not like it. The only thing that
I actually do to PATH is to add $HOME/.local/bin to it, so it is
easier to just add that to exec-path directly.