Use slime-lisp-implementations instead of inferior-lisp-program.
inferior-lisp-program does not seem to play well with use-package. It
doesn't work if I try to set it in (use-package slime),
(use-package emacs), or even (use-package inf-lisp). It does work if
I set it in a top-level setq, but I don't want that.
slime-lisp-implementations is a better way of choosing which Common
Lisp implementation to use, and it takes precedence.