Demand package.el use-package declaration.
Apparently, when I set use-package-always defer (ref: a35a113), I
introduced a bug where my package.el config would not be loaded
properly (as it was deferred). This only manifests on a fresh git
clone of my config where packages are not available yet, so it took
some time to notice, as I do not need to clone my config often.